requirements.txt
azure-communication-messages
fastapi
httpx
python-dotenv
uvicorn
websocket-client